# src/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Numgas API Contract
  version: 1.0.0
  description: >-
    Kontrak API untuk sistem Numgas meliputi Auth & Account, Public Access,
    SPPG Dashboard, dan School Dashboard.
servers:
  - url: /api/v1
    description: Default API base path
...
paths:
...
  /sppg/dashboard/summary:
    get:
      tags:
        - 03 - SPPG Dashboard
      summary: Dashboard Summary
      operationId: getSppgDashboardSummary
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Ringkasan dashboard SPPG berhasil diambil.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/SppgDashboardSummary'
                required:
                  - data
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Unauthorized:
      description: Tidak terautentikasi atau token tidak valid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    SppgDashboardSummary:
      type: object
      properties:
        sppgName:
          type: string
        operationalStatus:
          type: string
          enum: [active, inactive, maintenance]
        totalPartnerSchools:
          type: integer
        totalMenus:
          type: integer
      required:
        - sppgName
        - operationalStatus
        - totalPartnerSchools
        - totalMenus
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        details:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
      required:
        - message